+++ Journal/PlacesTableViewController.swift
@@ -54,7 +54,12 @@ class PlacesTableViewController: UITableViewController {
   
   override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
     let cell = tableView.dequeueReusableCell(withIdentifier: "PlaceCell", for: indexPath)
-    let location = LocationsStorage.shared.locations[indexPath.row]
+    let locations = LocationsStorage.shared.locations
+    guard indexPath.row < locations.count else {
+      return cell
+    }
+
+    let location = locations[indexPath.row]
     cell.textLabel?.numberOfLines = 3
     cell.textLabel?.text = location.description
     cell.detailTextLabel?.text = location.dateString